requirements.txt
torch
hydra-core
omegaconf
einops
tensordict
tqdm
